Update @orbit/indexeddb to use new query expressions
