+++ packages/@orbit/indexeddb/test/source-test.ts
@@ -12,7 +12,6 @@ import {
   removeFromHasMany,
   replaceHasMany,
   replaceHasOne,
-  oqb,
   Transform,
   Schema,
   Source
@@ -437,7 +436,7 @@ module('IndexedDBSource', function(hooks) {
       addRecord(jupiter),
       addRecord(io)
     ]))
-      .then(() => source.pull(oqb.records()))
+      .then(() => source.pull(q => q.findRecords()))
       .then(transforms => {
         assert.equal(transforms.length, 1, 'one transform returned');
         assert.deepEqual(
@@ -482,7 +481,7 @@ module('IndexedDBSource', function(hooks) {
       addRecord(jupiter),
       addRecord(io)
     ]))
-      .then(() => source.pull(oqb.records('planet')))
+      .then(() => source.pull(q => q.findRecords('planet')))
       .then(transforms => {
         assert.equal(transforms.length, 1, 'one transform returned');
         assert.deepEqual(
@@ -531,7 +530,7 @@ module('IndexedDBSource', function(hooks) {
           addRecord(io)
         ]));
       })
-      .then(() => source.pull(oqb.record(jupiter)))
+      .then(() => source.pull(q => q.findRecord(jupiter)))
       .then(transforms => {
         assert.equal(transforms.length, 1, 'one transform returned');
         assert.deepEqual(
